requirements.txt
python-telegram-bot[job-queue]
python-dotenv
pandas
yfinance
joblib
openai
numpy